+++ python/qualitylib/metric_source/__init__.py
@@ -52,4 +52,5 @@ from .release_candidates import ReleaseCandidates
 from .sonar import Sonar
 from .trello import TrelloBoard, TrelloActionsBoard, TrelloRiskBoard, TrelloUnreachableException
 from .wiki import Wiki
+from .zap_scan_report import ZAPScanReport