+++ python/qualitylib/requirement/product_quality.py
@@ -23,7 +23,8 @@ from .. import metric
 OWASP = Requirement(
     name='OWASP Top 10',
     url='https://www.owasp.org/',
-    metric_classes={metric.HighPriorityOWASPDependencyWarnings, metric.NormalPriorityOWASPDependencyWarnings})
+    metric_classes={metric.HighPriorityOWASPDependencyWarnings, metric.NormalPriorityOWASPDependencyWarnings,
+                    metric.HighRiskZAPScanAlertsMetric, metric.MediumRiskZAPScanAlertsMetric})
 
 TRUSTED_PRODUCT_MAINTAINABILITY = Requirement(
     name='Trusted Product Maintainability, version 6.1',